+++ src/Facades/RelationRepository.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Katalam\OnOfficeAdapter\Facades;
 
 use Illuminate\Support\Facades\Facade;
+use Katalam\OnOfficeAdapter\Facades\Testing\RelationRepositoryFake;
 use Katalam\OnOfficeAdapter\Query\RelationBuilder;
 
 /**
@@ -14,6 +15,13 @@ use Katalam\OnOfficeAdapter\Query\RelationBuilder;
  */
 class RelationRepository extends Facade
 {
+    public static function fake(array ...$fakeResponses): RelationRepositoryFake
+    {
+        static::swap($fake = new RelationRepositoryFake(...$fakeResponses));
+
+        return $fake;
+    }
+
     protected static function getFacadeAccessor(): string
     {
         return \Katalam\OnOfficeAdapter\Repositories\RelationRepository::class;