+++ src/Facades/Testing/RecordFactories/RelationFactory.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Katalam\OnOfficeAdapter\Facades\Testing\RecordFactories;
+
+class RelationFactory extends BaseFactory
+{
+    public string $type = 'idsfromrelation';
+}